fix(investment): 대시보드 탭에 단타/비교 메뉴 노출 + 새 전략 인라인 렌더링
[버그 1] 대시보드 → 투자 탭(/dashboard?tab=investment)의 SUB_TABS에서
단타·전략비교 메뉴가 누락되어 있었음.
- InvestmentTab의 SubTab 유니온에 'daytrading' | 'compare' 추가
- SUB_TABS 배열에 ⚡ 단타(분봉), 🔀 전략 비교 항목 추가

[버그 2] "새 전략" 버튼이 <Link href="/investment/strategy/new">로 외부 페이지
이동을 일으켜 대시보드 컨텍스트를 빠져나갔음. 사이드바·헤더가 사라지는
"별도 페이지" 인상.
- creatingStrategy state로 BacktestPanel 패턴과 동일하게 인라인 표시
- 저장 성공 시 onSaved 콜백 → 인라인 종료 + 전략 목록 새로고침
- 취소(ArrowLeft) → 인라인 종료
- /investment/strategy 독립 페이지에서도 동일 패턴 적용

[리팩토링] 페이지 콘텐츠를 컴포넌트로 추출하여 인라인 렌더 가능화:
- src/components/Investment/StrategyBuilder/StrategyBuilder.tsx (신규)
- src/components/Investment/DayTradingContent.tsx (신규)
- src/components/Investment/CompareContent.tsx (신규)
- 기존 페이지 파일은 7줄짜리 thin wrapper로 변환 (직접 URL 진입 유지)

검증 (production build + Playwright):
- /dashboard?tab=investment 진입 시 7개 탭 모두 노출
- 새 전략 클릭 → URL 변경 없이 빌더 인라인 표시 (사이드바 유지)
- 단타·비교 탭 클릭 시 콘텐츠 영역에서 정상 렌더링
- 빌드 통과
